requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
bcrypt
apscheduler
python-dotenv
psycopg2-binary
pydantic[email]
aiosmtplib
requests
# redis
